requirements.txt
certifi==2019.3.9
chardet==3.0.4
googlemaps==3.0.2
idna==2.8
mysql-connector-python==8.0.15
numpy==1.16.2
pandas==0.24.2
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
xlrd==1.2.0
XlsxWriter==1.1.6
